CRUD Update

+++ Blogg/views.py
@@ -2,7 +2,7 @@ from django.shortcuts import render,redirect
 from .models import BlogPost
 from .forms import PostForm
 
-def index(request):
+def homepage(request):
     posts = BlogPost.objects.all()
     return render(request,'Blogg/index.html',{'posts':posts})
 
@@ -16,8 +16,9 @@ def create_post(request):
 
 def delete_post(request, post_id):
     posts = BlogPost.objects.get(pk=post_id)
-    posts.delete()
-    return redirect('index')
+    if request.method == 'POST':
+        posts.delete()
+        return redirect('homepage')
 
 def update_post(request):
     pass
